Make toES5 available as a utility.
